Fold Channel and CompletionQueue from grpc_impl to grpc
Reverts: https://github.com/grpc/grpc/pull/19067
