--- include/grpcpp/completion_queue_impl.h
@@ -1,24 +1,0 @@
-/*
- *
- * Copyright 2019 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-#ifndef GRPCPP_COMPLETION_QUEUE_IMPL_H
-#define GRPCPP_COMPLETION_QUEUE_IMPL_H
-
-#include <grpcpp/impl/codegen/completion_queue_impl.h>
-
-#endif  // GRPCPP_COMPLETION_QUEUE_IMPL_H
