+++ include/grpcpp/impl/codegen/client_interceptor.h
@@ -28,11 +28,12 @@
 
 namespace grpc_impl {
 
-class Channel;
 class ClientContext;
 }  // namespace grpc_impl
 
 namespace grpc {
+
+class Channel;
 
 namespace internal {
 class InterceptorBatchMethodsImpl;
